+++ src/Aws/CloudFront/Resources/cloudfront-2013-11-22.php
@@ -15,7 +15,7 @@
  */
 
 return array (
-    'apiVersion' => '2013-11-11',
+    'apiVersion' => '2013-11-22',
     'endpointPrefix' => 'cloudfront',
     'serviceFullName' => 'Amazon CloudFront',
     'serviceAbbreviation' => 'CloudFront',
@@ -68,7 +68,7 @@ return array (
     'operations' => array(
         'CreateCloudFrontOriginAccessIdentity' => array(
             'httpMethod' => 'POST',
-            'uri' => '/2013-11-11/origin-access-identity/cloudfront',
+            'uri' => '/2013-11-22/origin-access-identity/cloudfront',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
             'responseClass' => 'CreateCloudFrontOriginAccessIdentityResult',
             'responseType' => 'model',
@@ -76,7 +76,7 @@ return array (
                 'xmlRoot' => array(
                     'name' => 'CloudFrontOriginAccessIdentityConfig',
                     'namespaces' => array(
-                        'http://cloudfront.amazonaws.com/doc/2013-11-11/',
+                        'http://cloudfront.amazonaws.com/doc/2013-11-22/',
                     ),
                 ),
             ),
@@ -121,7 +121,7 @@ return array (
         ),
         'CreateDistribution' => array(
             'httpMethod' => 'POST',
-            'uri' => '/2013-11-11/distribution',
+            'uri' => '/2013-11-22/distribution',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
             'responseClass' => 'CreateDistributionResult',
             'responseType' => 'model',
@@ -129,7 +129,7 @@ return array (
                 'xmlRoot' => array(
                     'name' => 'DistributionConfig',
                     'namespaces' => array(
-                        'http://cloudfront.amazonaws.com/doc/2013-11-11/',
+                        'http://cloudfront.amazonaws.com/doc/2013-11-22/',
                     ),
                 ),
             ),
@@ -309,6 +309,10 @@ return array (
                                 ),
                             ),
                         ),
+                        'SmoothStreaming' => array(
+                            'type' => 'boolean',
+                            'format' => 'boolean-string',
+                        ),
                     ),
                 ),
                 'CacheBehaviors' => array(
@@ -416,6 +420,10 @@ return array (
                                                 ),
                                             ),
                                         ),
+                                    ),
+                                    'SmoothStreaming' => array(
+                                        'type' => 'boolean',
+                                        'format' => 'boolean-string',
                                     ),
                                 ),
                             ),
@@ -643,7 +651,7 @@ return array (
         ),
         'CreateInvalidation' => array(
             'httpMethod' => 'POST',
-            'uri' => '/2013-11-11/distribution/{DistributionId}/invalidation',
+            'uri' => '/2013-11-22/distribution/{DistributionId}/invalidation',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
             'responseClass' => 'CreateInvalidationResult',
             'responseType' => 'model',
@@ -651,7 +659,7 @@ return array (
                 'xmlRoot' => array(
                     'name' => 'InvalidationBatch',
                     'namespaces' => array(
-                        'http://cloudfront.amazonaws.com/doc/2013-11-11/',
+                        'http://cloudfront.amazonaws.com/doc/2013-11-22/',
                     ),
                 ),
             ),
@@ -721,7 +729,7 @@ return array (
         ),
         'CreateStreamingDistribution' => array(
             'httpMethod' => 'POST',
-            'uri' => '/2013-11-11/streaming-distribution',
+            'uri' => '/2013-11-22/streaming-distribution',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
             'responseClass' => 'CreateStreamingDistributionResult',
             'responseType' => 'model',
@@ -729,7 +737,7 @@ return array (
                 'xmlRoot' => array(
                     'name' => 'StreamingDistributionConfig',
                     'namespaces' => array(
-                        'http://cloudfront.amazonaws.com/doc/2013-11-11/',
+                        'http://cloudfront.amazonaws.com/doc/2013-11-22/',
                     ),
                 ),
             ),
@@ -886,9 +894,9 @@ return array (
         ),
         'DeleteCloudFrontOriginAccessIdentity' => array(
             'httpMethod' => 'DELETE',
-            'uri' => '/2013-11-11/origin-access-identity/cloudfront/{Id}',
+            'uri' => '/2013-11-22/origin-access-identity/cloudfront/{Id}',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
-            'responseClass' => 'DeleteCloudFrontOriginAccessIdentity2013_11_11Output',
+            'responseClass' => 'DeleteCloudFrontOriginAccessIdentity2013_11_22Output',
             'responseType' => 'model',
             'parameters' => array(
                 'Id' => array(
@@ -926,9 +934,9 @@ return array (
         ),
         'DeleteDistribution' => array(
             'httpMethod' => 'DELETE',
-            'uri' => '/2013-11-11/distribution/{Id}',
+            'uri' => '/2013-11-22/distribution/{Id}',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
-            'responseClass' => 'DeleteDistribution2013_11_11Output',
+            'responseClass' => 'DeleteDistribution2013_11_22Output',
             'responseType' => 'model',
             'parameters' => array(
                 'Id' => array(
@@ -966,9 +974,9 @@ return array (
         ),
         'DeleteStreamingDistribution' => array(
             'httpMethod' => 'DELETE',
-            'uri' => '/2013-11-11/streaming-distribution/{Id}',
+            'uri' => '/2013-11-22/streaming-distribution/{Id}',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
-            'responseClass' => 'DeleteStreamingDistribution2013_11_11Output',
+            'responseClass' => 'DeleteStreamingDistribution2013_11_22Output',
             'responseType' => 'model',
             'parameters' => array(
                 'Id' => array(
@@ -1006,7 +1014,7 @@ return array (
         ),
         'GetCloudFrontOriginAccessIdentity' => array(
             'httpMethod' => 'GET',
-            'uri' => '/2013-11-11/origin-access-identity/cloudfront/{Id}',
+            'uri' => '/2013-11-22/origin-access-identity/cloudfront/{Id}',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
             'responseClass' => 'GetCloudFrontOriginAccessIdentityResult',
             'responseType' => 'model',
@@ -1034,7 +1042,7 @@ return array (
         ),
         'GetCloudFrontOriginAccessIdentityConfig' => array(
             'httpMethod' => 'GET',
-            'uri' => '/2013-11-11/origin-access-identity/cloudfront/{Id}/config',
+            'uri' => '/2013-11-22/origin-access-identity/cloudfront/{Id}/config',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
             'responseClass' => 'GetCloudFrontOriginAccessIdentityConfigResult',
             'responseType' => 'model',
@@ -1062,7 +1070,7 @@ return array (
         ),
         'GetDistribution' => array(
             'httpMethod' => 'GET',
-            'uri' => '/2013-11-11/distribution/{Id}',
+            'uri' => '/2013-11-22/distribution/{Id}',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
             'responseClass' => 'GetDistributionResult',
             'responseType' => 'model',
@@ -1090,7 +1098,7 @@ return array (
         ),
         'GetDistributionConfig' => array(
             'httpMethod' => 'GET',
-            'uri' => '/2013-11-11/distribution/{Id}/config',
+            'uri' => '/2013-11-22/distribution/{Id}/config',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
             'responseClass' => 'GetDistributionConfigResult',
             'responseType' => 'model',
@@ -1118,7 +1126,7 @@ return array (
         ),
         'GetInvalidation' => array(
             'httpMethod' => 'GET',
-            'uri' => '/2013-11-11/distribution/{DistributionId}/invalidation/{Id}',
+            'uri' => '/2013-11-22/distribution/{DistributionId}/invalidation/{Id}',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
             'responseClass' => 'GetInvalidationResult',
             'responseType' => 'model',
@@ -1155,7 +1163,7 @@ return array (
         ),
         'GetStreamingDistribution' => array(
             'httpMethod' => 'GET',
-            'uri' => '/2013-11-11/streaming-distribution/{Id}',
+            'uri' => '/2013-11-22/streaming-distribution/{Id}',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
             'responseClass' => 'GetStreamingDistributionResult',
             'responseType' => 'model',
@@ -1183,7 +1191,7 @@ return array (
         ),
         'GetStreamingDistributionConfig' => array(
             'httpMethod' => 'GET',
-            'uri' => '/2013-11-11/streaming-distribution/{Id}/config',
+            'uri' => '/2013-11-22/streaming-distribution/{Id}/config',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
             'responseClass' => 'GetStreamingDistributionConfigResult',
             'responseType' => 'model',
@@ -1211,7 +1219,7 @@ return array (
         ),
         'ListCloudFrontOriginAccessIdentities' => array(
             'httpMethod' => 'GET',
-            'uri' => '/2013-11-11/origin-access-identity/cloudfront',
+            'uri' => '/2013-11-22/origin-access-identity/cloudfront',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
             'responseClass' => 'ListCloudFrontOriginAccessIdentitiesResult',
             'responseType' => 'model',
@@ -1238,7 +1246,7 @@ return array (
         ),
         'ListDistributions' => array(
             'httpMethod' => 'GET',
-            'uri' => '/2013-11-11/distribution',
+            'uri' => '/2013-11-22/distribution',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
             'responseClass' => 'ListDistributionsResult',
             'responseType' => 'model',
@@ -1265,7 +1273,7 @@ return array (
         ),
         'ListInvalidations' => array(
             'httpMethod' => 'GET',
-            'uri' => '/2013-11-11/distribution/{DistributionId}/invalidation',
+            'uri' => '/2013-11-22/distribution/{DistributionId}/invalidation',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
             'responseClass' => 'ListInvalidationsResult',
             'responseType' => 'model',
@@ -1305,7 +1313,7 @@ return array (
         ),
         'ListStreamingDistributions' => array(
             'httpMethod' => 'GET',
-            'uri' => '/2013-11-11/streaming-distribution',
+            'uri' => '/2013-11-22/streaming-distribution',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
             'responseClass' => 'ListStreamingDistributionsResult',
             'responseType' => 'model',
@@ -1332,7 +1340,7 @@ return array (
         ),
         'UpdateCloudFrontOriginAccessIdentity' => array(
             'httpMethod' => 'PUT',
-            'uri' => '/2013-11-11/origin-access-identity/cloudfront/{Id}/config',
+            'uri' => '/2013-11-22/origin-access-identity/cloudfront/{Id}/config',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
             'responseClass' => 'UpdateCloudFrontOriginAccessIdentityResult',
             'responseType' => 'model',
@@ -1340,7 +1348,7 @@ return array (
                 'xmlRoot' => array(
                     'name' => 'CloudFrontOriginAccessIdentityConfig',
                     'namespaces' => array(
-                        'http://cloudfront.amazonaws.com/doc/2013-11-11/',
+                        'http://cloudfront.amazonaws.com/doc/2013-11-22/',
                     ),
                 ),
             ),
@@ -1407,7 +1415,7 @@ return array (
         ),
         'UpdateDistribution' => array(
             'httpMethod' => 'PUT',
-            'uri' => '/2013-11-11/distribution/{Id}/config',
+            'uri' => '/2013-11-22/distribution/{Id}/config',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
             'responseClass' => 'UpdateDistributionResult',
             'responseType' => 'model',
@@ -1415,7 +1423,7 @@ return array (
                 'xmlRoot' => array(
                     'name' => 'DistributionConfig',
                     'namespaces' => array(
-                        'http://cloudfront.amazonaws.com/doc/2013-11-11/',
+                        'http://cloudfront.amazonaws.com/doc/2013-11-22/',
                     ),
                 ),
             ),
@@ -1595,6 +1603,10 @@ return array (
                                 ),
                             ),
                         ),
+                        'SmoothStreaming' => array(
+                            'type' => 'boolean',
+                            'format' => 'boolean-string',
+                        ),
                     ),
                 ),
                 'CacheBehaviors' => array(
@@ -1702,6 +1714,10 @@ return array (
                                                 ),
                                             ),
                                         ),
+                                    ),
+                                    'SmoothStreaming' => array(
+                                        'type' => 'boolean',
+                                        'format' => 'boolean-string',
                                     ),
                                 ),
                             ),
@@ -1943,7 +1959,7 @@ return array (
         ),
         'UpdateStreamingDistribution' => array(
             'httpMethod' => 'PUT',
-            'uri' => '/2013-11-11/streaming-distribution/{Id}/config',
+            'uri' => '/2013-11-22/streaming-distribution/{Id}/config',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
             'responseClass' => 'UpdateStreamingDistributionResult',
             'responseType' => 'model',
@@ -1951,7 +1967,7 @@ return array (
                 'xmlRoot' => array(
                     'name' => 'StreamingDistributionConfig',
                     'namespaces' => array(
-                        'http://cloudfront.amazonaws.com/doc/2013-11-11/',
+                        'http://cloudfront.amazonaws.com/doc/2013-11-22/',
                     ),
                 ),
             ),
@@ -2377,6 +2393,9 @@ return array (
                                         ),
                                     ),
                                 ),
+                                'SmoothStreaming' => array(
+                                    'type' => 'boolean',
+                                ),
                             ),
                         ),
                         'CacheBehaviors' => array(
@@ -2470,6 +2489,9 @@ return array (
                                                         ),
                                                     ),
                                                 ),
+                                            ),
+                                            'SmoothStreaming' => array(
+                                                'type' => 'boolean',
                                             ),
                                         ),
                                     ),
@@ -2787,7 +2809,7 @@ return array (
                 ),
             ),
         ),
-        'DeleteCloudFrontOriginAccessIdentity2013_11_11Output' => array(
+        'DeleteCloudFrontOriginAccessIdentity2013_11_22Output' => array(
             'type' => 'object',
             'additionalProperties' => true,
             'properties' => array(
@@ -2797,7 +2819,7 @@ return array (
                 ),
             ),
         ),
-        'DeleteDistribution2013_11_11Output' => array(
+        'DeleteDistribution2013_11_22Output' => array(
             'type' => 'object',
             'additionalProperties' => true,
             'properties' => array(
@@ -2807,7 +2829,7 @@ return array (
                 ),
             ),
         ),
-        'DeleteStreamingDistribution2013_11_11Output' => array(
+        'DeleteStreamingDistribution2013_11_22Output' => array(
             'type' => 'object',
             'additionalProperties' => true,
             'properties' => array(
@@ -3089,6 +3111,9 @@ return array (
                                         ),
                                     ),
                                 ),
+                                'SmoothStreaming' => array(
+                                    'type' => 'boolean',
+                                ),
                             ),
                         ),
                         'CacheBehaviors' => array(
@@ -3182,6 +3207,9 @@ return array (
                                                         ),
                                                     ),
                                                 ),
+                                            ),
+                                            'SmoothStreaming' => array(
+                                                'type' => 'boolean',
                                             ),
                                         ),
                                     ),
@@ -3447,6 +3475,9 @@ return array (
                                 ),
                             ),
                         ),
+                        'SmoothStreaming' => array(
+                            'type' => 'boolean',
+                        ),
                     ),
                 ),
                 'CacheBehaviors' => array(
@@ -3541,6 +3572,9 @@ return array (
                                                 ),
                                             ),
                                         ),
+                                    ),
+                                    'SmoothStreaming' => array(
+                                        'type' => 'boolean',
                                     ),
                                 ),
                             ),
@@ -4180,6 +4214,9 @@ return array (
                                             ),
                                         ),
                                     ),
+                                    'SmoothStreaming' => array(
+                                        'type' => 'boolean',
+                                    ),
                                 ),
                             ),
                             'CacheBehaviors' => array(
@@ -4273,6 +4310,9 @@ return array (
                                                             ),
                                                         ),
                                                     ),
+                                                ),
+                                                'SmoothStreaming' => array(
+                                                    'type' => 'boolean',
                                                 ),
                                             ),
                                         ),
@@ -4770,6 +4810,9 @@ return array (
                                         ),
                                     ),
                                 ),
+                                'SmoothStreaming' => array(
+                                    'type' => 'boolean',
+                                ),
                             ),
                         ),
                         'CacheBehaviors' => array(
@@ -4863,6 +4906,9 @@ return array (
                                                         ),
                                                     ),
                                                 ),
+                                            ),
+                                            'SmoothStreaming' => array(
+                                                'type' => 'boolean',
                                             ),
                                         ),
                                     ),